Add handlers controllers and API for local
